JackieSharkSkin/forFigures/pyproject.toml
[project]
name = "forfigures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "marimo>=0.20.4",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "plotly>=6.6.0",
]
